+++ internal/query/config/options.go
@@ -80,6 +80,15 @@ func WithSessionIdleTimeToLive(idleTimeToLive time.Duration) Option {
 	}
 }
 
+// WithImplicitSessions is an option to execute queries using an implicit session
+// which allows the queries to be executed without explicitly creating a session.
+// Please note that requests with this option use a separate session pool.
+func WithImplicitSessions() Option {
+	return func(c *Config) {
+		c.implicitSession = true
+	}
+}
+
 func WithLazyTx(lazyTx bool) Option {
 	return func(c *Config) {
 		c.lazyTx = lazyTx