+++ internal/query/execute_query.go
@@ -32,7 +32,6 @@ type executeSettings interface {
 	ResourcePool() string
 	ResponsePartLimitSizeBytes() int64
 	Label() string
-	IsImplicitSession() bool
 }
 
 type executeScriptConfig interface {