+++ internal/query/session.go
@@ -36,13 +36,7 @@ func (s *Session) QueryResultSet(
 		onDone(finalErr)
 	}()
 
-	settings := options.ExecuteSettings(opts...)
-
-	if settings.IsImplicitSession() {
-		return nil, xerrors.WithStackTrace(errImplicitSessionsNotSupported)
-	}
-
-	r, err := s.execute(ctx, q, settings, withTrace(s.trace))
+	r, err := s.execute(ctx, q, options.ExecuteSettings(opts...), withTrace(s.trace))
 	if err != nil {
 		return nil, xerrors.WithStackTrace(err)
 	}
@@ -81,13 +75,7 @@ func (s *Session) QueryRow(ctx context.Context, q string, opts ...options.Execut
 		onDone(finalErr)
 	}()
 
-	settings := options.ExecuteSettings(opts...)
-
-	if settings.IsImplicitSession() {
-		return nil, xerrors.WithStackTrace(errImplicitSessionsNotSupported)
-	}
-
-	row, err := s.queryRow(ctx, q, settings, withTrace(s.trace))
+	row, err := s.queryRow(ctx, q, options.ExecuteSettings(opts...), withTrace(s.trace))
 	if err != nil {
 		return nil, xerrors.WithStackTrace(err)
 	}
@@ -176,10 +164,6 @@ func (s *Session) Exec(ctx context.Context, q string, opts ...options.Execute) (
 		onDone(finalErr)
 	}()
 
-	if settings.IsImplicitSession() {
-		return xerrors.WithStackTrace(errImplicitSessionsNotSupported)
-	}
-
 	r, err := s.execute(ctx, q, options.ExecuteSettings(opts...), withTrace(s.trace))
 	if err != nil {
 		return xerrors.WithStackTrace(err)
@@ -207,10 +191,6 @@ func (s *Session) Query(ctx context.Context, q string, opts ...options.Execute) 
 	defer func() {
 		onDone(finalErr)
 	}()
-
-	if settings.IsImplicitSession() {
-		return nil, xerrors.WithStackTrace(errImplicitSessionsNotSupported)
-	}
 
 	r, err := s.execute(ctx, q, options.ExecuteSettings(opts...), withTrace(s.trace))
 	if err != nil {