+++ internal/query/transaction.go
@@ -86,10 +86,6 @@ func (tx *Transaction) QueryResultSet(
 		onDone(finalErr)
 	}()
 
-	if txSettings.IsImplicitSession() {
-		return nil, xerrors.WithStackTrace(errImplicitSessionsNotSupported)
-	}
-
 	if tx.completed {
 		return nil, xerrors.WithStackTrace(errExecuteOnCompletedTx)
 	}
@@ -142,10 +138,6 @@ func (tx *Transaction) QueryRow(
 	defer func() {
 		onDone(finalErr)
 	}()
-
-	if txSettings.IsImplicitSession() {
-		return nil, xerrors.WithStackTrace(errImplicitSessionsNotSupported)
-	}
 
 	resultOpts := []resultOption{
 		withTrace(tx.s.trace),
@@ -210,10 +202,6 @@ func (tx *Transaction) Exec(ctx context.Context, q string, opts ...options.Execu
 	defer func() {
 		onDone(finalErr)
 	}()
-
-	if txSettings.IsImplicitSession() {
-		return xerrors.WithStackTrace(errImplicitSessionsNotSupported)
-	}
 
 	if tx.completed {
 		return xerrors.WithStackTrace(errExecuteOnCompletedTx)
@@ -287,10 +275,6 @@ func (tx *Transaction) Query(ctx context.Context, q string, opts ...options.Exec
 	defer func() {
 		onDone(finalErr)
 	}()
-
-	if txSettings.IsImplicitSession() {
-		return nil, xerrors.WithStackTrace(errImplicitSessionsNotSupported)
-	}
 
 	if tx.completed {
 		return nil, xerrors.WithStackTrace(errExecuteOnCompletedTx)