+++ query/execute_options.go
@@ -49,13 +49,6 @@ func WithExecMode(mode options.ExecMode) ExecuteOption {
 	return options.WithExecMode(mode)
 }
 
-// WithImplicitSession is an option to execute a query using an implicit session
-// which allows the query to be executed without explicitly creating a session.
-// Please note that requests with this option use a separate session pool.
-func WithImplicitSession() ExecuteOption {
-	return options.WithImplicitSession()
-}
-
 func WithSyntax(syntax options.Syntax) ExecuteOption {
 	return options.WithSyntax(syntax)
 }