+++ CHANGELOG.md
@@ -1,6 +1,5 @@
 * Added a feature: "Execute queries without explicit session creation":
-  * added `ydb.WithImplicitSessionPoolSizeLimit()` option for limitation of implicit sessions
-  * added `query.WithImplicitSession()` option for execute query with implicit session
+  * added `config.WithImplicitSessions` query option for execute queries with implicit sessions
 
 ## v3.112.0
 * Added support for the `json.Unmarshaler` interface in the `CastTo` function for use in scanners, such as the `ScanStruct` method